requirements.txt
Flask-JWT-Extended==4.6.0
Flask-Smorest==0.42.0
Flask-SQLAlchemy==3.1.1
passlib==1.7.4
marshmallow==3.21.1
python-dotenv==1.0.1
gunicorn==21.2.0
Flask-Migrate==4.0.5
psycopg2==2.9.9
requests==2.31.0
rq==1.15.1
redis==5.0.1
jinja2==3.1.2